Networked operation!
Split the graphical and the command-line utility. The command-line
utility can act as a server, the GUI utility can connect to it.
The GUI utility can also work standalone. But you'll have to run
it as root (as always, due to the permissions in /proc), so using
the command-line utility as backend is useful to avoid running the
GUI as root.